requirements.txt
chardet==5.2.0
pillow==11.2.1
qrcode==8.2
reportlab==4.4.1
